implement paper deletion

+++ PaperLinkBackend/src/main/java/com/louishhy/paperlinkbackend/service/PaperService.java
@@ -10,8 +10,11 @@ import com.louishhy.paperlinkbackend.repository.PaperRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Objects;
@@ -53,7 +56,11 @@ public class PaperService {
         return paperRepository.save(paper);
     }
 
+    @Transactional
     public void deletePaper(long id) {
+        if (!paperRepository.existsById(id)){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Paper not found");
+        }
         paperRepository.deleteById(id);
     }
